requirements.txt
Django==1.8.0
Pillow
MySQL-python
psycopg2
django-crispy-forms==1.4.0
django-contact-form==1.2